Refactored around a new metadata module
